Fix issues identified in PR review
