+++ src/routes/dataset.ts
@@ -42,8 +42,8 @@ import {
   updateSources,
   getDatasetById,
   deleteDraftDatasetById,
-  getEverythingFromDatalake,
-  listAllFilesInDataset
+  listAllFilesInDataset,
+  getAllFilesForDataset
 } from '../controllers/dataset';
 import { rateLimiter } from '../middleware/rate-limiter';
 
@@ -226,7 +226,7 @@ router.patch('/:dataset_id/topics', jsonParser, loadDataset(withDraftAndTopics),
 
 // GET /dataset/:dataset_id/download
 // Downloads everything from the datalake relating to this dataset as a zip file
-router.get('/:dataset_id/download', loadDataset(withDraftForCube), getEverythingFromDatalake);
+router.get('/:dataset_id/download', loadDataset(withDraftForCube), getAllFilesForDataset);
 
 // GET /dataset/:dataset_id/list-files
 // List all the files which are used to build the cube
