+++ src/utils/dataset-controller-utils.ts
@@ -1,0 +1,54 @@
+import JSZip from 'jszip';
+import { FileImportDto } from '../dtos/file-import';
+import { StorageService } from '../interfaces/storage-service';
+import { DataLakeFileEntry } from '../interfaces/datalake-file-entry';
+import { Dataset } from '../entities/dataset/dataset';
+import { FileImportType } from '../enums/file-import-type';
+
+export const addDirectoryToZip = async (
+  zip: JSZip,
+  datasetFiles: Map<string, FileImportDto>,
+  directory: string,
+  fileService: StorageService
+) => {
+  const directoryList = await fileService.listFiles(directory);
+  for (const fileEntry of directoryList) {
+    let filename: string;
+    if ((fileEntry as DataLakeFileEntry).name) {
+      const entry = fileEntry as DataLakeFileEntry;
+      if (entry.isDirectory) {
+        await addDirectoryToZip(zip, datasetFiles, `${directory}/${entry.name}`, fileService);
+        continue;
+      }
+      filename = (fileEntry as DataLakeFileEntry).name;
+    } else {
+      filename = fileEntry as string;
+    }
+    const originalFilename = datasetFiles.get(filename)?.filename || filename;
+    zip.file(originalFilename, await fileService.loadBuffer(filename, directory));
+  }
+};
+
+export const collectFiles = (dataset: Dataset): Map<string, FileImportDto> => {
+  const files = new Map<string, FileImportDto>();
+  if (dataset.measure.lookupTable) {
+    const fileImport = FileImportDto.fromFileImport(dataset.measure.lookupTable);
+    fileImport.type = FileImportType.Measure;
+    files.set(dataset.measure.lookupTable.filename, fileImport);
+  }
+  dataset.dimensions.forEach((dimension) => {
+    if (dimension.lookupTable) {
+      const fileImport = FileImportDto.fromFileImport(dimension.lookupTable);
+      fileImport.type = FileImportType.Dimension;
+      files.set(dimension.lookupTable.filename, fileImport);
+    }
+  });
+  dataset.revisions.forEach((revision) => {
+    if (revision.dataTable) {
+      const fileImport = FileImportDto.fromFileImport(revision.dataTable);
+      fileImport.type = FileImportType.DataTable;
+      files.set(revision.dataTable.filename, fileImport);
+    }
+  });
+  return files;
+};
